requirements.txt
astroid==1.5.3
isort==4.2.14
lazy-object-proxy==1.3.1
lxml==3.8.0
mccabe==0.6.1
pdfkit==0.6.1
pep8==1.7.0
pylint==1.7.1
python-slugify==1.2.4
six==1.10.0
Unidecode==0.4.20
urltools==0.3.2
wrapt==1.10.10
